maybe got the new dataset class made
